Extract staticmethods to a base class that can be used for testing & debugging unit tests & code
